pyproject.toml
[tool.poetry]
name = "llm4intent"
version = "0.1.0"
description = ""
authors = ["Command <me@c0mm4nd.com>"]
readme = "README.md"
packages = [
    { include = "LLM4Intent" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
autogen-agentchat = {extras = ["gemini"], version = "^0.2.40"}
python-dotenv = "1.0.1"
web3 = "6.20.3"
web3research = {git = "https://github.com/njublockchain/web3research-py"}
groq = {git = "https://github.com/groq/groq-python.git", rev = "generated"}
langchain-google-genai = "2.0.7"
langchain-google = "^0.1.1"
langgraph = "^0.2.60"
langchain-core = "^0.3.28"
langchain-community = "^0.3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
websockets = "^14.1"
langchain = "^0.3.15"
ipython = "^8.31.0"
langchain-anthropic = "^0.3.4"
langchain-groq = "^0.2.4"
transformers = "^4.48.3"
langchain-xai = "^0.2.1"
networkx = "^3.4.2"
poetry = "^2.1.1"
tavily-python = "^0.5.1"
neo4j = "^5.28.1"
neo4j-graphrag = {extras = ["openai"], version = "^1.5.0"}

[tool.poetry.scripts]
start = "LLM4Intent.__main__:start"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
